Test: RetrieveUsersForAGame

+++ PlaningPoker.Api/webapi/GameUsersReadDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace webapi;
 
 public record GameUsersReadDto(string Id, string CreatedBy, string Title, string Description,
-    int RoundTime, int Expiration, List<UsersReadDto> usersReadDtos);
+    long RoundTime, long Expiration, List<UsersReadDto> usersReadDtos);
